refactoring model class, pipeline class, added FP8 support
