+++ .gitignore
@@ -3,6 +3,9 @@ __pycache__/
 *$py.class
 /outputs/
 /input/
+/dataset/
+/save/
+/output/
 checkpoints/
 /.vs
 .vscode/
