+++ FaithDiff/create_FaithDiff_model.py
@@ -1,32 +1,34 @@
-from .pipelines.pipeline_FaithDiff_tlc import DiffaVA_lr_pipeline
+from utils.system import quantize_8bit
+from .pipelines.pipeline_FaithDiff_tlc import FaithDiffStableDiffusionXLPipeline
 import torch
 from diffusers import AutoencoderKL
-from .models.unet_2d_condition_w_vae import UNet2DConditionModel as UNet2DConditionModel_vae
+from .models.unet_2d_condition_vae_extension import UNet2DConditionModel
 from diffusers import AutoencoderKL, DDPMScheduler
 from .models.bsrnet_arch import RRDBNet as BSRNet
 
 
-def FaithDiff_pipeline(sdxl_path, VAE_FP16_path, FaithDiff_path):
-    vae = AutoencoderKL.from_pretrained(VAE_FP16_path).to(dtype=torch.float16)
-    unet = UNet2DConditionModel_vae.from_pretrained(sdxl_path, subfolder="unet", variant="fp16")
-    unet.init_vae_encoder()
-    unet.init_information_transformer_layes()
-    unet.init_ControlNetConditioningEmbedding()
-    unet.denoise_encoder.dtype=torch.float16
-    unet.load_state_dict(torch.load(FaithDiff_path, weights_only=False), strict=True)
-    unet = unet.to(dtype=torch.float16)
+def FaithDiff_pipeline(sdxl_path, VAE_FP16_path, FaithDiff_path, use_fp8 = False):
+    dtype = torch.float16
+    vae = AutoencoderKL.from_pretrained(VAE_FP16_path).to(dtype=dtype)
+    unet = UNet2DConditionModel.from_pretrained(sdxl_path, subfolder="unet", variant="fp16")
+    unet.load_additional_layers(weight_path=FaithDiff_path, dtype=dtype)    
+    if use_fp8:
+        quantize_8bit(unet)
+    else:
+        unet = unet.to(dtype=torch.float16)
+
     DDPM_scheduler = DDPMScheduler.from_pretrained(sdxl_path, subfolder="scheduler")
-    pipe = DiffaVA_lr_pipeline.from_pretrained(
+    pipe = FaithDiffStableDiffusionXLPipeline.from_pretrained(
         sdxl_path,
         vae = vae,
-        unet=unet,
-        add_sample = True,
+        add_sample = True,        
         denoise_encoder = unet.denoise_encoder,
         DDPM_scheduler = DDPM_scheduler,
         add_watermarker=False,
-        torch_dtype=torch.float16,
+        torch_dtype=dtype,
         variant="fp16"
     )
+    pipe.unet = unet
 
     return pipe
